requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
